added [[ ! -L "${i}" ]] - symlink check to create_filestructure()
since a common use is to symlinking a a config folder in order to backup it in many ways like git. I've added [[ ! -L "${i}" ]] - symlink check to create_filestructure() it will check if dir is missing or if it's not a symlink before trying to execute mkdir -p. 
since your only check was to check if its a Dir your install fails on mkdir -p since its already exists as symlink at my case.